+++ xclim/indicators/atmos/_precip.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from xclim import indices
-from xclim.core.indicator import Indicator
-from xclim.core.indicator import Indicator2D
+from xclim.core.indicator import Daily
+from xclim.core.indicator import Daily2D
 from xclim.core.utils import wrapped_partial
 
 __all__ = [
@@ -20,11 +20,12 @@ __all__ = [
 ]
 
 
-class Pr(Indicator):
+class Pr(Daily):
     context = "hydro"
 
 
-class PrTas(Indicator2D):
+class PrTas(Daily2D):
+    _nvar = 2
     context = "hydro"